pytorch_docker/requirements.txt
pandas==1.3.2
numpy==1.20.3
scikit-learn==0.24.2
matplotlib==3.4.3
google-cloud-storage==1.42.3